print_every_steps

+++ finetune_fsdp.py
@@ -507,8 +507,8 @@ if __name__ == "__main__":
 
                 metadata['total_optimizer_steps'] += 1
                 if args.print_every_steps and metadata['total_optimizer_steps'] % args.print_every_steps == 0:
-                    loss_numerator = torch.tensor([metadata['loss_numerator']], device=device)
-                    loss_denominator = torch.tensor([metadata['loss_denominator']], device=device)
+                    loss_numerator = torch.tensor([metadata['loss_numerator']], dtype=torch.float64, device=device)
+                    loss_denominator = torch.tensor([metadata['loss_denominator']], dtype=torch.float64, device=device)
                     torch.distributed.all_reduce_coalesced(
                         [loss_numerator, loss_denominator], op=torch.distributed.ReduceOp.SUM)
                     if rank == 0:
